Moved the dataContracts helper into a new subfolder.
