+++ source/services/dataContracts/dataService/data.service.ts
@@ -6,7 +6,7 @@ import * as _ from 'lodash';
 import { IArrayUtility, serviceName as arrayServiceName, moduleName as arrayModuleName } from '../../array/array.service';
 import { IBaseDataServiceBehavior, BaseDataServiceBehavior, IConverter, IGetListOptions } from '../baseDataServiceBehavior';
 import { IBaseResourceParams } from '../resourceBuilder/resourceBuilder.service';
-import { helper } from '../dataContractsHelper.service';
+import { helper } from '../dataContractsHelper/dataContractsHelper.service';
 
 export var moduleName: string = 'rl.utilities.services.dataContracts.dataService';
 export var factoryName: string = 'dataService';
